refactored Organizers tests to extrace featured image related functions
