+++ tests/functional/Tribe/Events/Importer/File_Importer_OrganizersTest.php
@@ -8,6 +8,13 @@ use Tribe__Events__Importer__File_Importer_Organizers as Organizers_Importer;
 
 class File_Importer_OrganizersTest extends \Codeception\TestCase\WPTestCase {
 
+	protected $field_map = [
+		'organizer_name',
+		'organizer_email',
+		'organizer_website',
+		'organizer_phone'
+	];
+
 	/**
 	 * @var \Tribe__Events__Importer__File_Reader
 	 */
@@ -43,9 +50,7 @@ class File_Importer_OrganizersTest extends \Codeception\TestCase\WPTestCase {
 		parent::setUp();
 
 		// your set up methods here
-		$this->handlebars              = new Handlebars( [
-			'loader' => new FilesystemLoader( codecept_data_dir( 'csv-import-test-files/featured-image' ) )
-		] );
+		$this->handlebars              = new Handlebars();
 		$this->featured_image_uploader = $this->prophesize( 'Tribe__Events__Importer__Featured_Image_Uploader' );
 	}
 
@@ -64,65 +69,30 @@ class File_Importer_OrganizersTest extends \Codeception\TestCase\WPTestCase {
 		$sut = $this->make_instance();
 	}
 
-	/**
-	 * @test
-	 * it should not mark record as invalid if featured image entry is missing
-	 */
-	public function it_should_not_mark_record_as_invalid_if_featured_image_entry_is_missing() {
-		$sut = $this->make_instance();
 
-		$post_id = $sut->import_next_row();
-
-		$this->assertNotFalse( $post_id );
-	}
-
-	/**
-	 * @test
-	 * it should import and attach featured image if featured image is ok
-	 */
-	public function it_should_import_and_attach_featured_image_if_featured_image_is_ok() {
-		$image_url     = plugins_url( '_data/csv-import-test-files/featured-image/images/featured-image.jpg', codecept_data_dir() );
-		$attachment_id = $this->factory()->attachment->create_upload_object( $image_url );
-		$this->featured_image_uploader->upload_and_get_attachment()->willReturn( $attachment_id );
-
-		$sut = $this->make_instance();
-
-		$post_id = $sut->import_next_row();
-
-		$this->assertEquals( $attachment_id, get_post_thumbnail_id( $post_id ) );
-	}
-
-	/**
-	 * @test
-	 * it should not import and attach featured image if featured image is not ok
-	 */
-	public function it_should_not_import_and_attach_featured_image_if_featured_image_is_not_ok() {
-		$this->featured_image_uploader->upload_and_get_attachment()->willReturn( false );
-
-		$sut = $this->make_instance();
-
-		$post_id = $sut->import_next_row();
-
-		$has_thumbnail = wp_get_attachment_url( get_post_thumbnail_id( $post_id ) );
-		$this->assertFalse( $has_thumbnail );
-	}
-
-	private function make_instance() {
-		$this->rendered_file_contents = $this->handlebars->loadTemplate( $this->template )->render( $this->data );
-		vfsStream::setup( 'csv_file_root', null, [ 'organizers.csv' => $this->rendered_file_contents ] );
-		$this->file_reader = new \Tribe__Events__Importer__File_Reader( vfsStream::url( 'csv_file_root/organizers.csv' ) );
-		$this->file_reader->set_row( 1 );
+	protected function make_instance( $template_dir = null ) {
+		$this->setup_file( $template_dir );
 
 		$sut = new Organizers_Importer( $this->file_reader, $this->featured_image_uploader->reveal() );
-		$sut->set_map( [
-			'organizer_name',
-			'organizer_email',
-			'organizer_website',
-			'organizer_phone'
-		] );
+		$sut->set_map( $this->field_map );
 		$sut->set_type( 'organizers' );
 
 		return $sut;
+	}
+
+	/**
+	 * @param $template_dir
+	 */
+	protected function setup_file( $template_dir ) {
+		if ( ! empty( $template_dir ) ) {
+			$this->handlebars->setLoader( new FilesystemLoader( codecept_data_dir( 'csv-import-test-files/' . $template_dir ) ) );
+			$this->rendered_file_contents = $this->handlebars->loadTemplate( $this->template )->render( $this->data );
+			vfsStream::setup( 'csv_file_root', null, [ 'venues.csv' => $this->rendered_file_contents ] );
+			$this->file_reader = new \Tribe__Events__Importer__File_Reader( vfsStream::url( 'csv_file_root/organizers.csv' ) );
+		} else {
+			$this->file_reader = new \Tribe__Events__Importer__File_Reader( codecept_data_dir( 'csv-import-test-files/organizers.csv' ) );
+		}
+		$this->file_reader->set_row( 1 );
 	}
 
 }
